Create A_Cute_Cryptography_Learning_Note_v20211223
The note is initially integrated from my notes of CS3230 and CS4236. The two modules help me construct the skeleton of this note, and thereafter I extended the contends based on Cryptography I on Coursera by Dan Boneh and some academic papers. Because the note was initially my learning note, I still do not include citations and references to some works that the note is based on. Those will be coming soon and the book will keep updating as I explore more on cryptography. Please contact me if there are any typos or mistakes in the proof and explanations. Moreover, welcome to contact me to be an editor of the learning notes together, which will be a means for us to progress.